VVV-02 | Possible integer overflow and dead loop

+++ contracts/Vaults/VaultVested.sol
@@ -289,7 +289,7 @@ contract VaultVested is IDistributable, ReentrancyGuard, DepositoryRestriction, 
     }
 
     function _deleteUser(address _account) private {
-        for (uint8 i = 0; i < users.length; i++) {
+        for (uint i = 0; i < users.length; i++) {
             if (users[i] == _account) {
                 delete users[i];
             }
